offline_tools/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.12.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
colour==0.1.5
contourpy==1.3.0
coverage==7.10.6
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.124.0
fonttools==4.59.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.2
kiwisolver==1.4.7
loguru==0.7.2
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.7.2
numpy==1.24.3
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.1
pathlib2==2.3.7.post1
Pillow==10.0.0
plotly==5.16.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
seaborn==0.12.2
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tomli==2.2.1
tqdm==4.66.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
webcolors==1.13
webdriver-manager==4.0.2
websocket-client==1.9.0
win32_setctime==1.2.0
wsproto==1.2.0
xlsxwriter==3.2.5
zipp==3.23.0
